State Machines
Using a task based approach, with a simple channel implementation,
the active LED and two buttons can be treated as state machines whose
state is evaluated on each loop. Based on the actions taken place and
the current state of the LED changes, the LED lights in the first row
of the LED matrix will change.